requirements.txt
pyyaml
readchar
requests
